Update how we default kadmind and KDC services
The way this cookbook originally would default the settings for
kadmind and krb5kdc KDC settings was suboptimal. Instead, we should
use Chef attributes correctly and assume the user is doing the right
thing when using the cookbook. For users running a single kadmind/KDC,
there should be no noticeable difference.

Signed-off-by: Chris Gianelloni <wolf31o2@gmail.com>