Convert LWRPs to custom resources
Convert the LWRP `krb5_principal` and `krb5_keytab` into Chef custom
resources. This is a breaking change and will require a major version
bump as it removes compatibility with older versions of Chef. This
introduces a new `host_keytab` recipe, which can be used to create a
host principal and add it to the keytab. This recipe is used to test the
custom resources in the new InSpec-based integration tests, added for
this recipe.

Signed-off-by: Chris Gianelloni <wolf31o2@gmail.com>